Switch to Entity classes
Using a primitive as entity in Java is a fool's errand, since every lookup
the entity is used for (its primary purpose) requires (auto)boxing to
object. We also never calculate with entity ids.

int seems rather limiting if you think about it, especially when you would
want to encode information in the entity id, like the machine it's coming
from or whatever.

So we're using our own type-safe Entity class with long payload and
equals/hashcode impl.