Add alive flag to entity
Of all schemes for an entity system client to get notified if an entity he
holds a reference to is still alive, an explicit set to to false on kill
is the easiest and most performant.

This also generalize org.javesy.util.SingletonSet as a set view on a single object.